+++ runtime/src/bank/pyth/mod.rs
@@ -1,0 +1,5 @@
+pub mod accumulator;
+mod batch_publish;
+
+#[cfg(test)]
+mod tests;
